+++ plugin-annotation/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
+### mapbox-android-plugin-annotation-v9:0.9.0 - XXX
+#### Bugs
+- Consumable events for click listeners [1124](https://github.com/mapbox/mapbox-plugins-android/pull/1124)
+- Support multiple manager to be draggable [1125](https://github.com/mapbox/mapbox-plugins-android/pull/1125)
+- End of drag event is consumed [1141](https://github.com/mapbox/mapbox-plugins-android/pull/1141)
+
 ### mapbox-android-plugin-annotation-v9:0.8.0 - March 5, 2020
 #### Features
 - Switching all plugins to AndroidX [#1100](https://github.com/mapbox/mapbox-plugins-android/pull/1100)